Cover missing cases during module extension label normalization
The previous logic missed to normalize cases such as `"extension.bzl"`
and `"//extension.bzl"`, thus resulting in crashes if these styles are
mixed as well as invalid buildozer commands for `use_repo` fixing.

Instead of enumerating cases, parse the label and emit it in unambiguous
canonical form with a leading `@` stripped.