python: Make bazel_tools register system Python toolchain
A regression from upgrading to rules_python 0.22.0 is that a Python
toolchain is no longer registered by default (the prior 0.4.0 version
registered the "autodetecting" toolchain). Subsequent versions of
rules_python (0.23.0 or later) register a toolchain by default again,
but it's a downloaded runtime, which makes upgrading more involved.

Instead, make the bazel_tools module register the "autodetecting"
toolchain to restore the rules_python 0.4.0 behavior. This should
suffice bazel_tools is upgraded to 0.23.0 or later.