Remove storage of the set of requested `ArtifactNestedSetKey` deps and use an `ImmutableSet` for input dep keys.
The purpose of `requestedArtifactNestedSetKeys` was to essentially intern those keys, but now that `ArtifactNestedSetKey` is interned (https://github.com/bazelbuild/bazel/commit/5bbd8c7edf47f4dd44bdb01d8fa972051516d451), this is a waste. Note that this was attempted previously in unknown commit (2020), but that was rolled back due to interner overhead, which was solved by[]

The second part of the change is to materialize input dep keys in an `ImmutableSet` instead of using a lazy `Iterable`. This will help with my "take 2" for b/314117358. Making this in the same change because without `requestedArtifactNestedSetKeys`, it's easier to see that a lazy `Iterable` isn't saving too much. Note that the set doesn't escape input accumulation, so it's eligible for GC by the time the action actually executes.

PiperOrigin-RevId: 590580512
Change-Id: Ib196f766a8d1676a820544f337872728e7a3b347