Extend deletion of incompletely rewound keys to include keys that were rewound and completed unsuccessfully.
Errors after rewinding are rare and imply flakiness, since the node was previously done and successful at the same version. To keep the graph in a consistent state, delete the reverse transitive closure of such nodes. It does not make sense to have a graph where a successful node depends on an error that it usually cannot recover from.

PiperOrigin-RevId: 590672702
Change-Id: I0f5b5c8033c7eff4284bfdfc70e4ab5e7209792d