Starlarkify `getBuildVariables`.
Unfortunately there is no good way of Starlarkifying this method without moving some of the techdebt from Java to Starlark. There are two root causes for this problem:
* [getCppConfigurationEvenThoughItCanBeDifferentThanWhatTargetHas](https://cs.opensource.google/bazel/bazel/+/master:src/main/java/com/google/devtools/build/lib/rules/cpp/CcToolchainProvider.java;l=724-739;drc=b074ddb86eb0cdb7ca96110d6c3c8b602dde08d8) I believe this can be deleted once [incompatible_disable_starlark_host_transitions](https://cs.opensource.google/bazel/bazel/+/master:src/main/java/com/google/devtools/build/lib/packages/semantics/BuildLanguageOptions.java;l=622-631;drc=106aec622f0cd011467bdc63b50982afc961e2d5) is set to `True` and removed.
* Right now because we have different build variables for apple and non-apple platforms we have to live with this ugly callback closure [functions](https://cs.opensource.google/bazel/bazel/+/master:src/main/starlark/builtins_bzl/common/cc/cc_toolchain_provider_helper.bzl;l=234-237;drc=14c8411279f8b462c2c100b64454f7b68880b14d) which calculate them. Either we have to remove this difference or remove support of build variables altogether in the future.

Also since there is Starlark as well as Java code referring to `getBuildVariables` this cl could have been simplified if we had left Java implementation in place so that Java code calls into Java implementation and Starlark calls into Starlark one. However this I believe introduces more pain and confusion for the future so I've decided to go against this decision.

PiperOrigin-RevId: 590866747
Change-Id: Ibff225ddf37684ece54c4df2171e79f20377cb6a