Remove the ability to use FilesToRunProvider in coverage_support_files .
The functionality was added in https://github.com/bazelbuild/bazel/commit/062b83fcf48206aba254607c701bdae5c928de0b and it never worked: it only added the runfiles of the tool to the actions as a runfiles middleman, but not into the RunfilesSupplier, which means that they never materialized onthe input of these actions.

RELNOTES[INC]: Passing a FilesToRunProvider to coverage_support_files will now result in an error as opposed to being silently ignored.

PiperOrigin-RevId: 590926984
Change-Id: I72d2f0d4fdff8c4cf5951c640e93cec25fb936c7