python: make incompatible_python_disallow_native_rules work for top-level external repo targets
This basically makes it usable with the downloaded runtimes rules_python makes
available, which reference their runtimes as e.g. `@python_3_11//:python`.

The issue was the code to construct the label was leaving a trailing "/" when
the the target being checked at the root of the workspace. To fix, just omit
the trailing slash when the package name is empty to prevent the trailing
slash.

Work towards https://github.com/bazelbuild/bazel/issues/17773

Closes https://github.com/bazelbuild/bazel/pull/20516

PiperOrigin-RevId: 591053422
Change-Id: I7790df2db10278844ae2b36cfe671de03164972f