[6.5.0] RemoteSpawnRunner: record inbetween phases in timing profile (#20550)
After an action was executed remotely, RemoteSpawnRunner would use the
timestamps in the execution metadata to record appropriate timing phases
into the JSON profile.

However, there are durations in-between the existing phases that are
unaccounted for. Depending on the RBE server implemenation, these phases
could mean different things:
- Sandbox preparation
- Cleaning up sandbox environments post-execution
- Others

Missing these durations inside the timing profile would cause confusion
to end users as it would be interpreted as nothing happened in between
the existing phases.

Add these durations into the profile as "pre-X" phases so that user is
aware of activities could still be happening during that time. RBE
server implementation should be able to alter these label
programmatically if necessary.

Closes #20387.

Commit
https://github.com/bazelbuild/bazel/commit/fe9e9e04cbc7174fc33fa18810ff6647766100c4

PiperOrigin-RevId: 590816782
Change-Id: I2bee36be928db24a14fab18bc519c3893723b7d6

Co-authored-by: Son Luong Ngoc <sluongng@gmail.com>