Split off various fields of Package into a new Package.Metadata object
A Metadata consists only of information that can be determined by the end of evaluating a BUILD file, but not any symbolic macros it may contain. The idea is to use Metadata objects wherever the full Package is not required -- basically anywhere we're not iterating over the targets. This can enable lazy symbolic macro evaluation, so that we don't run macros for targets that are not requested.

Work toward #19922.

PiperOrigin-RevId: 591271790
Change-Id: I28916eef9cef66fb4ea52ef985963e233044b9cd