Remove the distinction for removing "old" vs "new" rdeps, which was only used for preconditions checks.
Callers were responsible for calling either `removeReverseDep()` (old) or `removeInProgressReverseDep()` (new). In certain circumstances, rewinding makes it impossible for the caller to know whether the rdep is old or new with the information that it has. Merge the two into a single `removeReveseDep()` method that works regardless of old/new and loosen the precondition checks accordingly.

Consider the case where `top` depends on `dep`, and both are inflight. Without rewinding, we have the invariant that when `dep` completes, it must always signal `top`. With rewinding, however, it is possible that `dep` was previously done and already signaled `top`. In this scenario, the rdep would have been consolidated (moved to the "old" set) when `dep` had its value set.

PiperOrigin-RevId: 591911288
Change-Id: Ifd1c594ddecdaebed5657658461609c207da57cb