Remove/rewrite 'Arrange'/'Act'/'Assert' test comments that added no details.
The 'Arrange/Act/Assert' testing style is popular inside the Bazel codebase,
and for particularly complex test cases it can be helpful to mark the sections
of the test while explaining the details of each test's 'stage'. (This is
particularly useful when testing concurrent behaviors where 'Arrange' must
prepare concurrency primitives with exact blocking flow control, and 'Act' will
be coupled to the concurrency primitives used in 'Arrange').

It is not however considered helpful to include plain comments labeling the
'stage' of each test without any other detail. These comments can be either
deleted (if the test case is understandable without the labels) or rewritten to
add details that were originally elided.

PiperOrigin-RevId: 591943992
Change-Id: I537b6dfa96f860389790c2b7d4bbe4001cbf1535