Initial commit for `bazel focus`.
The `focus` command is an experimental feature to minimize retained heap usage in Bazel. It works by accepting a set of files, known as the "working set", and then deterministically prunes unnecessary Skyframe nodes for correct incremental builds for this set.

Early prototypes have observed a 90% reduction of post-build, post-gc retained heap with a single file in the working set of a large server-based project.

This change has no actual focus logic yet. It adds the skeleton code and basic tests for the new command. The main focus logic is expressed as pseudocode, details in follow up changes.

PiperOrigin-RevId: 592488344
Change-Id: I1d195e17998044c22fa0dacfdb0dd035e0afc407