Refactor `FdoHelper` so that it no longer depends on `CcToolchainAttributesProvider`.
**Why This is necessary?**

With the removal of `incompatible_enable_cc_toolchain_resolution` there is no need for `CcToolchainAttributesProvider` to exist. The only remaining place from where it is referenced is `FdoHelper` (apart from `CcStarlarkInternal` however this is not the blocker for deletion). Therefore to make removal process possible I've decided to get rid of this dependency.

**Are we adding more tech debt?**

No. I plan to remove `CcToolchainAttributesProvider` in the near future(~1 week). This cl just simplifies that removal process a lot for both future me and a future reviewer. Otherwise I had to merge this refactoring into the removal CL which would become huge.

PiperOrigin-RevId: 592508211
Change-Id: I7b0a7acd932b63d578ca4bb6486f30d1ff678790