python: rules_python 0.22.0 -> 0.22.1 soas to register Python toolchain by default
A regression from upgrading to rules_python 0.22.0 is that a Python toolchain
is no longer registered by default (the prior 0.4.0 version registered the
"autodetecting" toolchain). Subsequent versions of rules_python (0.23.0 or
later) register a toolchain by default again, but it's a downloaded runtime,
which makes upgrading more involved.

To fix, upgrade to 0.22.1, which registers the autodetecting toolchain.
This restores the rules_python 0.4.0 behavior and should suffice until
bazel_tools is upgraded to 0.23.0 or later.

Fixes https://github.com/bazelbuild/bazel/issues/20458

Closes https://github.com/bazelbuild/bazel/pull/20514

PiperOrigin-RevId: 592578175
Change-Id: I707a4c96829063536f81cc10e144652102da1e7e