Move more fields into Metadata
Each of the three affected fields is initialized when the builder is constructed, so they may as well go alongside the other fields that are determined prior to BUILD evaluation. The affected fields are also conceivably useful to symbolic macros as well, so they probably belong on Metadata for that reason too.

Work toward #19922.

PiperOrigin-RevId: 592602808
Change-Id: Ia171bb67ff6d386cc9687ad46e05e8c6aeade1f8