Add `GlobFunctionWithStateMachine` implementation and apply it to most of the `SkyframeExecutor`s.
## Problem with canonical recursive implementation:
The canonical `GlobFunction` implementation recursively creates sub-glob nodes until the full glob pattern is evaluated. If the glob pattern contains many fragments and/or the package contains multiple sub-directories from vertical and/or horizontal dimensions. The sub-glob nodes tree can potentially become giant, leading to significant memory overhead.

This commit intends to eliminate the potentially giant sub-glob nodes tree for cases when incrementality performance is not strict.

## Work done in this commit:
* Refactor `GlobFunction` and `GlobValue` into two pairs of subclasses:
  * `GlobFunctionWithMultipleRecursiveFunctions` and `GlobValueWithNestedSet`: This pair is a copy of the canonical implementation of `GlobFunction`. It is still preferred to be used when incrementality performance is strict.
  * `GlobFunctionWithRecursionInSingleFunction` and `GlobValueWithImmutableSet`: This pair is the new `GlobFunction` implementation using Skyframe StateMachine mechanism (https://bazel.build/contribute/statemachine-guide).
* Implement multiple `StateMachine` producers under `third_party/bazel/src/main/java/com/google/devtools/build/lib/packages/producers` directory to support `GlobFunctionWithRecursionInSingleFunction` computation;
* Switch to use `GlobFunctionWithRecursionInSingleFunction` for all conditions which does not require strict incremental evaluation performance;
* Refactor and add more tests in `GlobFunctionTest` to cover both recursive and State Machine globbing logic.

## Benefits of using GlobFunctionWithRecursionInSingleFunction:
* `GlobFunctionWithRecursionInSingleFunction` uses `StateMachine`s as the recursive globbing units instead of Glob nodes. The `StateMachine`s are discarded after each blaze build or query invocation. And thus, it eliminates giant sub-Glob nodes tree cached in Skyframe.
* The number of skyframe restarting times is also significantly reduced. This is due to `StateMachine` enables structured concurrency when querying dependent `SkyKey`s.

## Limitations of current `GlobFunctionWithRecursionInSingleFunction`:
Current `GlobFunctionWithRecursionInSingleFunction` is not performance friendly for incremental blaze evaluation. Due to the lose of the sub-glob nodes tree, skyframe incrementality is totally lost. So the computation cost of clean and non-clean glob evaluation is the same with `GlobFunctionWithRecursionInSingleFunction`, even when this is just a simple filesystem change in the package.

On the other hand, with `GlobFunctionWithMultipleRecursiveFunctions`, the sub-Glob nodes tree cached in skyframe guarantees incrementality. A simple filesystem change in the package leads to re-evaluation a very small subset of the sub-Glob nodes tree.

Experiments have also shown significant performance regression for certain incremental evaluations.

PiperOrigin-RevId: 592650084
Change-Id: Iff4096ca2ebb14bd3045c06056ba4bad34dd396b