Modify the error message that occurs when a requested target does not exist.
First, we removed the suggestion to "use `query`" as it was misleading to some users.

Second, we fine-tuned the suggested target behaviors. The existing method retrieves only the closest target within a reasonable distance. We updated this to return potentially more targets, with a preference for those that are particularly close.

Performance-wise, this should be not diverge much from the status quo. The existing behavior already checks edit-distance for each target in the package against the requested target. This adds minimal behavior on top of that to identify the most likely candidate(s).

RELNOTES: None.
PiperOrigin-RevId: 591302621
Change-Id: I9bd272b68def2f9a75db01061287697d23936bca

Cherry-picked from ef98ef99b2ff8d177891b1b093295b11913921d6