Move PackageContext fields into Package.Builder
This is to prepare for eliminating the PackageContext class itself, in favor of Package.Builder.

PackageContext's getLabel() and setMakeVariable() accessors are replaced by the corresponding methods of the underlying Builder.

The globber and eventHandler fields are migrated to the Builder as well. Since we don't have a PackageContext in all circumstances (e.g., when a Package.Builder is instantiated for bzlmod logic or unit tests), these fields are @Nullable, and the caller must beware. The eventHandler field in particular is just a direct migration from PackageContext; a follow-up CL will further refactor it so that eventHandler is not set by the caller but rather owned outright by the builder, and can never be null.

Minor notes:
- Since globber is now stored on the builder, it no longer needs to be threaded as a separate arg to executeBuildFile[Impl]().
- Callers who would've read the eventHandler off the PackageContext now retrieve it from the Builder, so we now need to assign the eventHandler to the Builder via the new setLocalEventHandler() in PackageFactory and WorkspaceFactory. Again, this is temporary.

Continued from unknown commit.

Work toward #19922.

PiperOrigin-RevId: 594123106
Change-Id: I2bd898c57755a4d7bd687aff526305168478ce94