Change LocalResourcesSupplier to ResourceSetOrBuilder
The objective is to rewrite CppLinkAction to Starlark. In Starlark only ResourceSetOrBuilder is available.

Drop computation of inputsBytes. The size of inputs is not available is Starlark. This regresses logging and any potential data collection based on the logs. Unfortunately Starlarkification of CppLinkAction will need to completely remove such type of logging. Those logs were introduced recently. They are released with Bazel 7.0.0, so the support period of Bazel 7 should give enough time to collect the data.

PiperOrigin-RevId: 594376814
Change-Id: I5f8cb4ee09a51db420881bbbc9941e746990bbe5