Remove linkerInputArtifacts from LinkCommandLine
They value of linkerInputArtifacts was only used in tests.

Replace uses in tests with CppLinkCommand.getInputs(). The second method also returns path to linker and some scripts used during linker, so exact comparisons need to be replaced with containsAtLeast. It's more reliable on one hand (because actions can't lie about their inputs), and less reliable (because there could be some additional files on the path that are weren't there before). However, general intent of the tests is verifying that the files are present.

PiperOrigin-RevId: 594431021
Change-Id: I84cdda8d4bc50715ea3e412760bf8c713abd0353