Automated rollback of commit 132a5f96d08b616ae9cda636eced1b7317a751e9.
*** Reason for rollback ***

Going forward, the FDO type SAFDO will no longer be used to distinguish regular and configless AutoFDO.  Both flavours will use FDO type AFDO.

*** Original change description ***

Have getFdoBuildStamp return "SAFDO" when a .afdo profile is supplied along with xbinary_fdo.

This change enables the fdo_type build stamp to be set to SAFDO in order to distinguish and track the prevalence of SAFDO in the fleet. As an alternative to this implementation, I also considered defining an SAFDO build feature and using a new file extention to identify SAFDO file. This CL is here unknown commit. I decided this implementation is preferable since from a blaze perspective, SAFDO is really...

***

PiperOrigin-RevId: 595128289
Change-Id: Ie1bc2f1a7dc023ec42965cfee477a391434863bb