Propagate command line args (if any) to the underlying bazel invocation in `update_default_lock_file.sh`
Motivation is to allow running (some) Bazel tests on a `rules_java` presubmit. This is tricky because there is no simple way to override a dep in `MODULE.tools`. Plan is to use a fake registry and pass the appropriate `--registry=` flag(s). So, this requires propagating the flags while updating the tests' lock file.

PiperOrigin-RevId: 595339606
Change-Id: Ib1858eab29a7013a08ab1bebddf954e2f2f0a47a