Push some Package.Builder settings into the constructor
This makes the generatorMap, configSettingVisibilityPolicy, and globber into Builder() constructor args, instead of setter mutations. These kinds of settings are always known before BUILD file evaluation begins, so it's best to bake this guarantee into the builder's API. This makes it slightly easier to follow the rest of the builder's logic.

setFilename() will be done in a follow-up. We may or may not push setLoads() up in this manner.

This does make the builder's constructor signature longer, but I think that's preferable to mixing initial setup mutations with actual package construction mutations. One way to address this might be to have the caller supply the Package.Metadata object to the builder constructor, but first we'd have to migrate off Metadata the fields that are determined through BUILD file evaluation.

Work toward #19922.

PiperOrigin-RevId: 595407792
Change-Id: I5ed2b07b631fdbbf393a2095acd342a6e53d3c25