Use long instead of int for thread indices
This protects against an integer overflow which could occur for large key list size and large thread counts.

Regrettably, it's difficult to write a regression test for this scenario, as exercising this overflow requires lots of time and heap, so it would be a performance regression to our test suites.

Fixes https://github.com/bazelbuild/bazel/issues/19445

PiperOrigin-RevId: 595420516
Change-Id: Ic0a475a6a273c50fe9895dd0852fa5b062859cb2