Eliminate package builder's setFilename method
Follow-up to unknown commit.

The initialization of the metadata fields is moved to the top of the builder constructor. The creation of the InputFile corresponding to the BUILD file is moved to the bottom of the builder constructor. Had to replace a use of Builder#createLabel with Label#create to avoid a circular init problem where pkg was not yet initialized.

Work toward #19922.

PiperOrigin-RevId: 595426821
Change-Id: I1e48dae33231d6ca3eb3feebbf9e202a1d1b2dec