Clarify name conflict checking invariants
Previously, it was possible in theory, but not in practice, to call both addRule() and addRuleUnchecked() on the same Package.Builder. This CL removes that possibility, and in doing so makes it easier to reason about whether the ruleLabels map will be null. It also makes the outputFilePrefixes map nullable along the same lines as ruleLabels. Both maps are only used for validation, and so are irrelevant to Builders that use addRuleUnchecked.

For completeness/correctness, replaceTarget is also updated to mention its interaction with this invariant.

Work toward #19922.

PiperOrigin-RevId: 595696336
Change-Id: I3965033784025e3dd6cfc85ec0b591a21c24de8c