[7.1.0] Accept labels of aliases in config_setting. (#20649)
Previously, if a config_setting referenced a Starlark setting through an
alias, it was always evaluated as if the setting had its default value.
Now, it is evaluated correctly.

This is done by looking up the value of the build setting in the
configuration based on its own label as specified in
BuildSettingProvider.

Fixes #13463 .

RELNOTES: None.
Commit
https://github.com/bazelbuild/bazel/commit/d44c3f98f717328d399899005e6e5ed1263e9725

PiperOrigin-RevId: 585658985
Change-Id: Id534cd95282355f1143302bf703145bb53708a41

Co-authored-by: Googler <lberki@google.com>
Co-authored-by: Ian (Hee) Cha <heec@google.com>