Automated rollback of commit 13bd828893be702f960027b83e9b2e4649a0bd37.
*** Reason for rollback ***

Broke tests

*** Original change description ***

Update CompatDexBuilder to compile using the DexFilePerClassFileConsumer.

The builder is already compiling a single classfile to a single dex file (in DexIndexed mode). The use of the DexFilePerClassFileConsumer mode will not change that, but it informs D8 that the compilation is an "intermediate" compilation and more closely matches the usage.

PiperOrigin-RevId: 596000561
Change-Id: If44a3c63264cdbb482e0de5809ee2e6479b5c614