[7.1.0] Cover missing cases during module extension label normalization (#20630)
The previous logic missed to normalize cases such as `"extension.bzl"`
and `"//extension.bzl"`, thus resulting in crashes if these styles are
mixed as well as invalid buildozer commands for `use_repo` fixing.

Instead of enumerating cases, parse the label and emit it in unambiguous
canonical form with a leading `@` stripped.

Closes #20482.

Commit
https://github.com/bazelbuild/bazel/commit/71787cf0552ac5e9b4ae458c8b3137cbe33b35a9

PiperOrigin-RevId: 592666970
Change-Id: Ieea34b27a187545a11107a334bbae14fef974ae8

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>
Co-authored-by: Ian (Hee) Cha <heec@google.com>
Co-authored-by: Xùdōng Yáng <wyverald@gmail.com>