[7.1.0] Remove unnecessary `cc_test` coverage handling (#20641)
Both the `lcov_merger` and the `collect_cc_coverage` script are obtained
from "well-known" implicit attributes of Starlark rules and do not
require any further special handling.

Removing this special handling also fixes a fringe bug: When a `cc_test`
is wrapped in another test target that applies a transition to it, the
removed logic set the `LCOV_MERGER` variable to the untransitioned path
of the `lcov_merger`, which will then not be available in the test
sandbox since it is only added as a runfile, but the test action
references it via its exec path. This resulted in a "file not found"
error that broke coverage.

Closes #20349.

Commit
https://github.com/bazelbuild/bazel/commit/591d1257c51c2f997e067964878e3a9c7ee1f89c

PiperOrigin-RevId: 589252093
Change-Id: I810f1b414dcfedcc8930a74390ef5c9bfd2cd030

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>
Co-authored-by: Ian (Hee) Cha <heec@google.com>