Circuit Breaker for HTTP Cache Client
This PR addresses an issue discovered during the evaluation of fallback HTTP remote cache support. It was noted that the circuit breaker functioned effectively with gRPC, but not for HTTP, due to the absence of Failure Circuit Breaker integration in the RemoteRetrier init context for HTTP.

To resolve this, the PR implements the integration of the same retrier used for gRPC into the HTTP context as well.

Closes #20831.

PiperOrigin-RevId: 597195032
Change-Id: Ib970023fa889ed3ce3c241dd794c9ad573ec61c6