Remove RunfilesSupport.getRunfiles().
Its only call site was used to merge the runfiles of --run_under= targets into the runfiles of other binaries, but it didn't work as desired: runfiles middlemen in runfiles trees are not expanded recursively.

Since filesToRun is just filesToBuild + the runfiles middleman and the runfiles middleman is ignored, the new code is equivalent to the old one.

RELNOTES: None.
PiperOrigin-RevId: 597251730
Change-Id: I6ae717cd0c1f31e1b42ab87b37e93162301000ae