[7.1.0] Introduce a SpawnLogContext interface. (#20842)
For now, the sole implementation is ExpandedSpawnLogContext, which
subsumes the old SpawnLogContext class, as well as some of the logic
around log formats currently in SpawnLogModule.

In a followup, a separate CompactSpawnLogContext implementation will be
introduced. This will let us experiment with a new log format while
minimizing the chance of accidentally breaking the existing formats.

PiperOrigin-RevId: 588444807
Change-Id: I75bc2a577ec45202baf95c950a257eaf420cbeb0