C++: Fix versioned shared libraries for macOS
The toolchain was passing -l:name. This mechanism doesn' t exist on macOS,
instead the full path to the shared library should be passed.

To add a proper test like in https://github.com/bazelbuild/bazel/pull/20847,
this change must first be checked in, the Apple toolchain needs to be
updated and then the test can be added.

RELNOTES:none
PiperOrigin-RevId: 597366229
Change-Id: Ib7cf237cb6c58a69e781197408d3f703010909da