When a TreeFileArtifact is on the inputs of an action, add the expansion of its parent to the input metadata map (not just its state)
The lack of this caused breakage in a very specific case: when an action had a TreeFileArtifact on its inputs (but not its parent) and its parent was discovered as an input, the expansion of the parent wouldn't be present in the ArtifactExpander and thus would be missing from the inputs of the sandboxed/remoted action.

RELNOTES: None.
PiperOrigin-RevId: 597522977
Change-Id: Ia5967c9a88fb8f4fa3d5f5ae725feac3d734cf17