Put a reference to RunfilesTree into MiddlemanAction and simplify the creation of the latter.
The plan is that it will be plumbed to RunfilesArtifactValue, thereby making RunfilesSupplier unnecessary.

I don't know yet how that will work; the most likely approach is an ugly downcast of Action to MiddlemanAction in ArtifactFunction, but I hope I can come up with something more palatable. So for now, the private field remains unused.

Since now there is only one kind of middleman, it was possible to delete a pleasant amount of code, including MiddlemanActionTest.

RELNOTES: None.
PiperOrigin-RevId: 597526236
Change-Id: I392fc662288cc5dbee3a058d246182916343f8cc