[7.1.0] Ignore read-only errors when updating the `mtime` of the `install_base` (#20648)
Currently if the `--install_base` path passed is not writable by the
user invoking Bazel, the Bazel client crashes:
```console
❯ bazel --install_base=/some/read/only/path version
FATAL: failed to set timestamp on '/some/read/only/path': (error: 30): Read-only file system
```

This happens because the Bazel client (unconditionally) attempts to
update the `mtime` of this path:

https://github.com/bazelbuild/bazel/blob/a3c677dfea2de636a719d50345a5a97af96fae60/src/main/cpp/blaze.cc#L1010-L1021

This commit updates the client to ignore such errors. See #20373 for
context.

Closes #20442.

Commit
https://github.com/bazelbuild/bazel/commit/7f782e3c1a1c4a21beea19c2cedb614be8316e67

PiperOrigin-RevId: 591266054
Change-Id: If53e7cad48cb62406f7883f72b413e4b5a0bb8e2

Co-authored-by: Rahul Butani <rrbutani@users.noreply.github.com>
Co-authored-by: Ian (Hee) Cha <heec@google.com>