[7.1.0] Flip flag `--experimental_use_semaphore_for_jobs`. (#20646)
Bazel 7 upgrades its runtime JVM to 21. However, the implementation of
`ForkJoinPool` in JDK 21 is changed so that the number of concurrent
actions may be larger than `--jobs`.

Flag `--experimental_use_semaphore_for_jobs` was introduced to fix this
problem. This CL flips it to make `--jobs` work again by default.

Fixes #20521.

Commit
https://github.com/bazelbuild/bazel/commit/3c298fdb868d17d30ef5c19b3f3b488e5ed2c8a9

PiperOrigin-RevId: 590834204
Change-Id: I0ddf1e1a088eb4d917036350533df94c17ac48e6

Co-authored-by: Googler <chiwang@google.com>
Co-authored-by: Yun Peng <pcloudy@google.com>