Fix versioned shared libraries for macOS toolchain (#20847)
The toolchain was passing -l:name. This mechanism doesn' t exist on
macOS, instead the full path to the shared library should be passed.

Mainline commit:
https://github.com/bazelbuild/bazel/commit/f0ade80ce920be0719b1a43a40258397f68a944d

As it says on that mainline commit, it cannot be checked in with a test
unlike this PR, because the macOS toolchain was moved outside of Bazel.
Therefore, the change must go in, the macOS toolchain needs to be
updated, then the test can be added. This change does go in with a test
because the toolchain is still embedded in Bazel for 6.5.

Fixes https://github.com/bazelbuild/bazel/issues/20487