Move caching from SingleRunfilesSupplier to RunfilesTree.
This removes the last piece of functionality from RunfilesSupplier, thus paving the way for plumbing RunfilesTrees to action execution in through the data path used for plumbing all other kinds of artifact metadata.

(it also paves the way for caching non-test runfiles trees and thus some reduction  in garbage generated, but that's not why I'm doing it)

The strange duplication of sh_test and java_test in the newly added test case is useful because sh_test is the only test rule class that uses the non-Starlark code path.

Also, as a drive-by change, remove SingleRunfilesSupplier.withOverriddenRunfilesDir() (it was unused in production) and SingleRunfilesSupplierTest (SingleRunfilesSupplier is now a very thin shim around Runfiles so it doesn't make much sense to test it)

RELNOTES: None.
PiperOrigin-RevId: 598807682
Change-Id: I627336045923bd0e5f7e4666c68b7bf0401f9bcc