Wire-up default_linkstatic attribute of cc_test_toolchain.
The intention of this attribute is to allow cc_test toolchains to specify
whether they support dynamic_mode linking or not.

In essence, dynamic_mode for cc_test (the generation of shared libs
per-cc_library) is test-runner specific--e.g. this could work fine for cc_test
run under emulation, mounting runfiles, but would be cumbersome when running
on-device.

PiperOrigin-RevId: 598865391
Change-Id: Ibfde7dfbb8ff8637ce5ba7c6062093679c57f1fb