Automated rollback of commit c6ba1c3302b3a766544158f47fe9247d6df3f8d7.
*** Reason for rollback ***

Version 8.4.7-dev landed with unknown commit with a fix for the marker issue

*** Original change description ***

Automated rollback of commit 13bd828893be702f960027b83e9b2e4649a0bd37.

*** Reason for rollback ***

Broke tests

*** Original change description ***

Update CompatDexBuilder to compile using the DexFilePerClassFileConsumer.

The builder is already compiling a single classfile to a single dex file (in DexIndexed mode). The use of the DexFilePerClassFileConsumer mode will not change that, but it informs D8 that the compilation is an "intermediate" compilation and more closely matches the usage.

PiperOrigin-RevId: 599148994
Change-Id: I308e131a7b21933ec06adedfcef1e265884ea37c