Do not depend on PRECOMPUTED:repo_env if the repository doesn't depend on any environment variable.
HEAD is wasteful because sometimes just checking the up-to-dateness of fetched repositories takes a lot of time (if there is a lot of hashing to be done), so it's better to get a Skyframe cache hit if at all possible.

This is not the full answer because repositories that depend only on unchanged environment variables will still need to be re-checked if there is an environment variable that does change, but it's still a step forward.

RELNOTES: None.
PiperOrigin-RevId: 599242763
Change-Id: I895c5793ed06ef2c7a3337ef232ab13a7596b325